backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Flask-Cors==4.0.0
flake8==7.0.0
black==24.3.0 